refactor: change RadioButtonType import position
